minor modifs

+++ libft/ft_strindof.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-int		ft_strindof(const char *s, char c)
+size_t		ft_strindof(const char *s, char c)
 {
 	size_t	i;
