+++ libft/libft.h
@@ -40,7 +40,7 @@ void			*ft_memccpy(void *dst, const void *src, int c, size_t n);
 void			*ft_memmove(void *dst, const void *src, size_t len);
 char			**ft_split(char const *s, char c);
 int				ft_strcountch(const char *s, size_t n);
-int				ft_strindof(const char *s, char c);
+size_t			ft_strindof(const char *s, char c);
 size_t			ft_strlen(const char *s);
 size_t			ft_arrlen(char **s);
 char			*ft_strchr(const char *s, int c);
